🌱 split preflightCheck: extract checks into separate module
